+++ package/src/server/components/server_provider.tsx
@@ -1,7 +1,7 @@
 import { setLocaleCache, setMessageForLocaleCache } from "../../general/cache_variables.js";
 import { getMessage } from "../functions/server.js";
 import type { TranslationObject } from "../../types/types.js";
-import dynamic from "next/dynamic";
+import dynamic from "next/dynamic.js";
 import { localesSet } from "../../config/middleware.js";
 import config from "../../config/intl_config.js";
 import type { SerializedAuthUser } from "../../firebase_auth/types.js";
@@ -106,7 +106,7 @@ let authUserServerProviderModule: typeof AuthUserServerProviderModule | undefine
  */
 export default async function LocationzationProvider({ language, messages, staticSafe = true, children }: { language: string, messages?: TranslationObject, staticSafe?: boolean, children: React.ReactNode }): Promise<Component> {
     if (!localesSet.has(language)) {
-        const { notFound } = await import("next/navigation");
+        const { notFound } = await import("next/navigation.js");
         notFound();
     }